proj_1/requirements.txt
boto3>=1.34.0
requests>=2.31.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tavily-python>=0.3.0